+++ R/Set_UniversalSet.R
@@ -6,8 +6,9 @@
 #' @details
 #' The Universal set is the default universe to all sets, and is the largest possible set.
 #' The Universal set contains every single possible element. We denote the Universal set with `V`
-#' instead of `U` to avoid confusion with the union symbol. The Universal set is also responsible
-#' for a few set paradoxes, to resolve these we use the following results:
+#' instead of `U` to avoid confusion with the union symbol. The Universal set cardinality is set to
+#' `Inf` where we assume `Inf` is greater than any `Aleph` or `Beth` numbers. The Universal set is
+#' also responsible for a few set paradoxes, to resolve these we use the following results:
 #'
 #' Let \eqn{V} be the universal set, \eqn{S} be any non-universal set, and \eqn{0} the empty set, then
 #'